ingest_mysql/requirements.txt
boto3
python-dotenv
pandas
mysql-connector-python
